Introduces pyparsing based parser.
Simplifies parser and string handling.

This fixes issue #47 (Last field in a record is not parsed when it does
not have brackets nor comma). See additional test cases to
"article_missing_coma.bib".

Adds optional behavior for bibdata to recognize a set of common strings
such as month expansion. This is not handled by bparser yet.

Also changes tests to fix some behaviors:
- preambles contains string expressions, thus quotes are mandatory if
  content is not a string name previously defined. This is fixed in
  tests, in bwriter and in in the new parser.
- update tests not to use undefined strings such as *jan*.