Add script to use BibTex to split string into separate names.
A custom BibTex style file is used which prints the input author string
followed by a list of the names within that string, one per line. The
Python script collects names (either from the commandline or
interactively), and then runs BibTex to generate the output.

This is used to generate test cases for the splitnames() function.