Add utils.splitnames() to break string into list of individual names.
This splits the string in the same manner as BibTeX, i.e., ignoring
instances of the separator ' and ' at a brace level greater than zero.
It also treats leading or trailing 'and's as part of a name, again as
per BibTeX.

This is covered by a number of testcases. The misc/splitnames.py script
can be used to split strings with BibTeX to generate extra tests.