Fix `convert_to_unicode` for unbraced sequences
`convert_to_unicode` customization was converting `\"u` (without braces)
to `\u0308u` which was supposed to be an `u` letter with an umlaut.
However, the `\u0308` unicode character is supposed to be *after* the
letter, contrary to the LaTeX encoding of the accent.

This commit fixes it and additionally wrap the `convert_to_unicode`
customization in a dedicated function that could be reused.

Closes #121.