Handles declarations on lines starting with spaces after comments.
(See #198)

The parser currently skips declarations preceded by a space after a comment:
they are considered part of the implicit comment before. This behavior differs,
for example, from what biber would do on the same file. It is also not consistent
with the fact that a line started as " @string" is a valid string declaration.