Fix #275.
The issue was caused by pyparsing not returning as named result an empty
string. Hence, although these are correctly parsed, they cannot be
accessed by names. The rule would apply for any part of the code
accessing by name a parse result where the empty string is admissible:
in such a case, the combination of `get` with a default of `""` should
be used instead of the direct access ot the named value.