align_values allows int or bool now
Fixes #315

The `align_values` of the BibTexWriter now accepts a bool or int value.
If the bool value `true` is specified, the maximal number of characters used in any fieldname is used as length. If a integer value is specified, the greater of the specified
integer value and the overall maximal number of characters used in any
fieldname is used.

This commit also fixes the previous behaviour of align_values which
results in a breaking change. The
`ENTRYTYPE` entry was considered for calculating the maximal number of
characters, which always lead to a minimum value of `9`. Now, keys that
are not written into the BibTex output are ignored which leads to an
exact computation of the fieldname lengths.

Example:
```
@book{abc123,
 a         = {test},
 bb        = {longvalue}
}
@book{abc123,
 a  = {test},
 bb = {longvalue}
}
```