:sparkles: handle concatenated fields with inner quotes in splitter (#398)
This allows the splitter to correctly handle #-based string concatenation.

Note: This will still lead to downstream problems, as some of these concatenated fields will not have a recognized enclosing, and as string interpolation does not yet work with concatenated references. However, these cases did not work before either and this this PR does not (knowingly) introduce any regressions. The hereby mentioned problems will be addressed in a subsequent PR.

This is the first pr to address (but not yet close) #396